backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
flower==2.0.1

# 现有ShadowForge依赖
openai==1.6.1
pillow==10.1.0
pygments==2.17.2
opencv-python==4.9.0.80
numpy==1.26.2
gtts==2.4.0
edge-tts==6.1.9
pyttsx3==2.90
reportlab==4.0.4
python-docx==1.1.0
python-pptx==0.6.23
requests==2.31.0
httpx==0.25.2